handling errors and improved output
A little tweak to make this tool perfect.

Well, I saw that, even if the system isn't connected to any network, the tool kept running and while stopping the process, it showed errors which kept ruining the beauty of this tool and the code. ( check the screenshot )
Even the uneven gaps didn't look good.

![before](https://cloud.githubusercontent.com/assets/9411252/15578778/70e54bd8-237f-11e6-878a-44613663609d.png)

So, I tweaked it little to make it even more good, showing messages and handling errors in a better way ( You can compare the screenshots )

![after](https://cloud.githubusercontent.com/assets/9411252/15578985/52d3e28e-2380-11e6-936d-0cd15d163ff1.png)

Well, the downloading speed is quite low here..haha

Closes #8